Swap contact for social
Removed the contact form above the footer and replaced it with our social media twitter icon and link to https://twitter.com/MimirSolutions. Contact form is currently unnecessary and the submission form does not currently have a PHP(mail) function so all submission go nowhere. Eliminating this feature and focusing on the social media on the main page is the best option for connection.